Extract send message command to CommandSendMessage
