+++ backend/demo/src/main/java/app/k9mail/backend/demo/DemoHelper.kt
@@ -1,0 +1,7 @@
+package app.k9mail.backend.demo
+
+import java.util.UUID
+
+internal object DemoHelper {
+    fun createNewServerId() = UUID.randomUUID().toString()
+}
